Add files via upload
# Minimum Work Per Day to Finish Tasks

## Description

This project implements a solution to determine the minimum amount of work to be done per day to finish given tasks within a specified number of days. The problem is defined as follows:

Given an array `task[]` of size `N` denoting the amount of work to be done for each task, find the minimum amount of work to be done on each day so that all the tasks can be completed in at most `D` days. Note: On one day, work can be done for only one task.

## Example

### Input
- `task[] = [3, 4, 7, 15]`
- `D = 10`

### Output
- `4`

### Input
- `task[] = [30, 20, 22, 4, 21]`
- `D = 6`

### Output
- `22`

## Installation

1. Clone the repository:
    ```sh
    git clone https://github.com/yourusername/minimum-work-per-day.git
    cd minimum-work-per-day
    ```

2. Compile the code:
    ```sh
    g++ -o minimum_work main.cpp
    ```

## Usage

1. Run the executable:
    ```sh
    ./minimum_work
    ```

2. Follow the prompts to enter the number of tasks, the amount of work for each task, and the maximum number of days available.

## Contributing

Contributions are welcome! Please fork the repository and submit a pull request with your changes.

## License

This project is licensed under the MIT License. See the LICENSE file for details.

## Contact

For any questions or suggestions, please open an issue or contact me at pravinthorve1282004@gmail.com.